PCA_Discharge widget was causing unhandled exception error which crashed the webpage stopping all communications.
ALT PBB_PCA_GPU_Status_Widget is showing _OutOfService and PerfectHookup, it is not supposed too.
-  Changed tagconfig for ATL: PBB_PCA_GPU_Status_Widget System._OutOfService and System.PerfectHookup, set ParentID:null, class:'no_row'

Alarm Widget is still causing the other widgets to stop, OAS said it was probably because we were running out of callbacks still when we added the Alarm

Widget.
-  Changed max_callbacks and callback_timeout to include looking for opc.alarm_bindings, then multiplies alarm_bindings.count * 5

Added the Comm priority 999 Alarms, need a new color for those alarms.
-  Added comm errors (priority 999) to Alar CSS and made the row color black background and red text

It is taking too long to determine when widgets are not loading.  We have to wiat for data to change and know what we are looking for in order to know

that we have good communications.
-  Added a comm heartbeat to each widget so if 3 of the client refresh cycles occur without getting a data_update call the border will go red letting the

end user know that they have no communications.
- Reporting widget needed to have the evaluation of a tag to ensure the comm heartbeat worked.

Is there a way the end user can restart comm without supports help?
When Updating settings for widgets, every change triggers a call to update which makes all the widgets on the screen flash and slows down the process of

changing settings.
- In order to address the issue of having a way for the end user to restart a widget I first had to address the issue of the update routine triggering

with every setting change.  This was fixed by moving the IsUpdatingSettings flag from the update routine to the toggle_settings routine and cleaning up

the toggle_settings routine so it is responsible for turning off the updates and calling the update routine.  The update routine exits immediately if the

IsUpdatingSettings or IsPageLoading flags are set, thus preventing it from reloading the settings with every setting change.  When the user returns to the

display mode the IsUpdatingSettings is set to false and the update routine processes, thus loading the widget with the final settings.
- Fixing the settings issue and having the toggle_settings routine manage turning off and calling the update routine allows for a widget with no comm to

be restarted by simply having the user go into the settings and then going back into th display mode, thus restarting the comm polling.

Add Out Of Service and Perfect Hookup to PBB_PCA_GPU_Status_Widget
- Added the logic to use the System._OutOfService and System.PerfectHookup tags to show icons on the widget.  NOTE:  Tags still need to be created in OAS

for all clients for the icons to show.

All Widgets are not loading to default client
- Added missing check_init_site() to widgets that didn't have it.
- Report widget: Cleaned up the onshow, set_model and toggle_settings to make the widget load the settings for the default widget.

Reports Widget won't load if user is assigned to one site only.
- Fixing the loading of the report widget loading the default client also resolved this issue.

When a widget size is increased too far it collapses.
- This was a setting for the jQuery.GridSter which limited the size to 20 (20*70 wide and 20*45 tall).  I set this to false to allow for no max size.

Add popup/tooltip to dashboard icons when collapsed
- Added tooltip to dashboard icons

The dashboard title is not showing like it used too
- set z-index of content-header to 1000 and positioned it on the header to save realistate